LayoutTests/inspector/elements/elements-panel-styles.html

<html>
<head>
<link rel="stylesheet" type="text/css" href="resources/elements-panel-styles.css">

<script src="../../http/tests/inspector/inspector-test.js"></script>
<script src="../../http/tests/inspector/elements-test.js"></script>
<script>

function test()
{
    InspectorTest.selectNodeAndWaitForStylesWithComputed("foo", step1);

    function step1()
    {
        InspectorTest.dumpSelectedElementStyles(false, false);
        InspectorTest.completeTest();
    }
}

</script>
</head>

<body onload="runTest()">
<p>
Tests that elements panel shows proper styles in the sidebar panel.
</p>

<div id="container">
    <div id="foo" class="foo" style="display:none" align="left">Foo</div>
</div>

</body>
</html>
