LayoutTests/inspector/elements/elements-panel-styles-expected.txt

Tests that elements panel shows proper styles in the sidebar panel.

[expanded] 
border-bottom-left-radius: 5px;
    .foo - 5px elements-panel-styles.css:14
border-bottom-right-radius: 5px;
    .foo - 5px elements-panel-styles.css:14
border-top-left-radius: 5px;
    .foo - 5px elements-panel-styles.css:14
border-top-right-radius: 5px;
    .foo - 5px elements-panel-styles.css:14
color: blue;
    .foo, .foo::before - blue elements-panel-styles.css:21
    /-- overloaded --/ .foo - black elements-panel-styles.css:14
display: none;
    element.style - none 
    /-- overloaded --/ div - block user agent stylesheet
font-family: serif;
    #container - serif elements-panel-styles.css:5
font-size: 14px;
    #container - 14px elements-panel-styles.css:5
    /-- overloaded --/ body - 12px elements-panel-styles.css:1
font-style: italic;
    #container .foo - italic !important elements-panel-styles.css:10
    /-- overloaded --/ .foo - normal !important elements-panel-styles.css:14
margin-bottom: 2px;
    .foo - 2px elements-panel-styles.css:14
margin-left: 0px;
    .foo - 0px elements-panel-styles.css:14
margin-right: 0px;
    .foo - 0px elements-panel-styles.css:14
margin-top: 10px;
    .foo - 10px elements-panel-styles.css:14
text-align: -webkit-left;
    div[] - -webkit-left 

[expanded] 
element.style  { ()
display: none;

======== Matched CSS Rules ========
[expanded] 
#container .foo  { (elements-panel-styles.css:10)
font-style: italic !important;

[expanded] 
.foo  { (elements-panel-styles.css:39)

[expanded] 
.foo, .foo::before  { (elements-panel-styles.css:21)
content: "[before Foo]";
color: blue;

[expanded] 
.foo  { (elements-panel-styles.css:14)
/-- overloaded --/ color: black;
margin: 10px 0 2px;
    margin-top: 10px;
    margin-right: 0px;
    margin-bottom: 2px;
    margin-left: 0px;
border-radius: 5px;
    border-top-left-radius: 5px;
    border-top-right-radius: 5px;
    border-bottom-right-radius: 5px;
    border-bottom-left-radius: 5px;
/-- overloaded --/ font-style: normal !important;

[expanded] 
div[]  { ()
text-align: -webkit-left;

[expanded] 
div  { (user agent stylesheet)
/-- overloaded --/ display: block;

======== Inherited from div#container ========
[expanded] 
#container  { (elements-panel-styles.css:5)
font-family: serif;
font-size: 14px;

======== Inherited from body ========
[expanded] 
body  { (elements-panel-styles.css:1)
/-- overloaded --/ font-size: 12px;


======== Pseudo ::before element ========
[expanded] 
.foo::before  { (elements-panel-styles.css:43)

[expanded] 
.foo::before  { (elements-panel-styles.css:26)
color: red;

[expanded] 
.foo, .foo::before  { (elements-panel-styles.css:21)
content: "[before Foo]";
/-- overloaded --/ color: blue;


======== Pseudo ::after element ========
[expanded] 
.foo::after  { (elements-panel-styles.css:34)
font-family: courier;
content: "[after Foo 2]";

[expanded] 
.foo::after  { (elements-panel-styles.css:30)
/-- overloaded --/ content: "[after Foo]";
color: green;
